Restructured to MVC closely and Improved error reporting
